requirements.txt
selenium>=4.8.0
webdriver-manager>=4.0.0
pytest>=7.0
python-dotenv>=0.21.0

# Lint / formatting (dev)
ruff>=0.4.0
black>=24.1.0